@@ -25,12 +25,10 @@
     "start:styling": "yarn workspace @stream-io/video-styling run start",
     "build:client": "yarn workspace @stream-io/video-client run build",
     "start:client": "yarn workspace @stream-io/video-client run start",
-    "start:i18n": "yarn workspace @stream-io/i18n start",
-    "build:i18n": "yarn workspace @stream-io/i18n build",
-    "build:react:deps": "yarn build:client && yarn build:i18n && yarn build:styling && yarn build:react:bindings && yarn build:react:sdk",
-    "build:react-native:deps": "yarn build:client && yarn build:i18n && yarn build:react:bindings && yarn build:react-native:sdk",
+    "build:react:deps": "yarn build:client && yarn build:styling && yarn build:react:bindings && yarn build:react:sdk",
+    "build:react-native:deps": "yarn build:client && yarn build:react:bindings && yarn build:react-native:sdk",
     "build:vercel": "yarn build:react:deps && yarn build:react:dogfood",
-    "build:react:video-demo": "yarn build:react:deps && yarn workspace react-video-demo run build:website",
+    "build:react:video-demo": "yarn build:react:deps && yarn workspace @stream-io/react-video-demo run build:website",
     "start:egress": "yarn workspace @stream-io/egress-composite start",
     "build:egress": "yarn workspace @stream-io/egress-composite build",
     "build:all": "yarn workspaces foreach -v --topological-dev run build",
@@ -52,7 +50,6 @@
     "sample:react:tutorial": "yarn workspace @stream-io/stream-video-react-tutorial run dev",
     "start:docs:react": "yarn workspace @stream-io/video-react-sdk run start:docs ",
     "release": "nx run-many --target version --parallel=1",
-    "release:18n": "yarn workspace @stream-io/i18n npm publish --access=public",
     "release:react-bindings": "yarn workspace @stream-io/video-react-bindings npm publish --access=public",
     "release:react-sdk": "yarn workspace @stream-io/video-react-sdk npm publish --access=public",
     "release:react-native-sdk": "yarn workspace @stream-io/video-react-native-sdk npm publish --access=public"
@@ -62,20 +59,20 @@
     "@commitlint/config-angular": "^17.0.0",
     "@jscutlery/semver": "^2.30.1",
     "@nrwl/devkit": "^16.0.1",
-    "@types/eslint": "^8.4.8",
-    "@typescript-eslint/eslint-plugin": "^5.47.0",
-    "@typescript-eslint/parser": "^5.47.0",
-    "@typescript-eslint/typescript-estree": "^5.47.0",
-    "eslint": "^8.28.0",
+    "@types/eslint": "^8.44.4",
+    "@typescript-eslint/eslint-plugin": "^6.8.0",
+    "@typescript-eslint/parser": "^6.8.0",
+    "@typescript-eslint/typescript-estree": "^6.8.0",
+    "eslint": "^8.51.0",
     "eslint-config-react-app": "^7.0.1",
-    "eslint-import-resolver-typescript": "^3.5.2",
-    "eslint-plugin-import": "^2.26.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-prettier": "^4.2.1",
     "ngx-deploy-npm": "^5.2.0",
     "nx": "16.0.1",
-    "prettier": "^2.8.4",
-    "typescript": "^4.9.5",
-    "vercel": "^32.1.0",
-    "vite": "^4.4.9"
+    "prettier": "^2.8.8",
+    "typescript": "^5.2.2",
+    "vercel": "^32.4.1",
+    "vite": "^4.4.11"
   }
 }
