@@ -1,6 +1,6 @@
 {
   "name": "@stream-io/video-client",
-  "version": "0.3.32",
+  "version": "0.3.35",
   "packageManager": "yarn@3.2.4",
   "main": "dist/index.cjs.js",
   "module": "dist/index.es.js",
@@ -9,16 +9,14 @@
   "types": "dist/index.d.ts",
   "license": "See license in LICENSE",
   "scripts": {
-    "postinstall": "yarn copy-version",
     "clean": "rimraf dist",
-    "start": "yarn copy-version && rollup -w -c",
-    "build": "yarn clean && yarn copy-version && rollup -c",
+    "start": "rollup -w -c",
+    "build": "yarn clean && rollup -c",
     "test": "vitest",
     "clean:docs": "rimraf generated-docs",
     "test-ci": "vitest --coverage",
     "generate:open-api": "./generate-openapi.sh protocol",
-    "generate:open-api:dev": "./generate-openapi.sh chat",
-    "copy-version": "echo \"export const version = '$npm_package_version';\" > ./version.ts"
+    "generate:open-api:dev": "./generate-openapi.sh chat"
   },
   "files": [
     "dist",
@@ -33,7 +31,7 @@
     "@protobuf-ts/runtime": "^2.9.1",
     "@protobuf-ts/runtime-rpc": "^2.9.1",
     "@protobuf-ts/twirp-transport": "^2.9.1",
-    "@types/ws": "^8.5.6",
+    "@types/ws": "^8.5.7",
     "axios": "^1.5.1",
     "base64-js": "^1.5.1",
     "isomorphic-ws": "^5.0.0",
@@ -46,19 +44,19 @@
   },
   "devDependencies": {
     "@openapitools/openapi-generator-cli": "^2.7.0",
-    "@rollup/plugin-replace": "^5.0.2",
-    "@rollup/plugin-typescript": "^11.1.2",
+    "@rollup/plugin-replace": "^5.0.4",
+    "@rollup/plugin-typescript": "^11.1.5",
     "@types/jsonwebtoken": "^9.0.3",
     "@types/sdp-transform": "^2.4.7",
     "@types/ua-parser-js": "^0.7.37",
     "@vitest/coverage-v8": "^0.34.4",
     "dotenv": "^16.3.1",
     "happy-dom": "^11.0.2",
-    "prettier": "^2.8.4",
-    "rimraf": "^5.0.1",
-    "rollup": "^3.28.1",
-    "typescript": "^4.9.5",
-    "vite": "^4.4.9",
+    "prettier": "^2.8.8",
+    "rimraf": "^5.0.5",
+    "rollup": "^3.29.4",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.11",
     "vitest": "^0.34.4",
     "vitest-mock-extended": "^1.2.1"
   }
