@@ -1,6 +1,6 @@
 {
   "name": "@stream-io/video-react-native-sdk",
-  "version": "0.1.6",
+  "version": "0.1.10",
   "packageManager": "yarn@3.2.4",
   "main": "dist/commonjs/index.js",
   "module": "dist/module/index.js",
@@ -45,7 +45,6 @@
     "!**/.*"
   ],
   "dependencies": {
-    "@stream-io/i18n": "workspace:^",
     "@stream-io/video-client": "workspace:^",
     "@stream-io/video-react-bindings": "workspace:^",
     "intl-pluralrules": "2.0.1",
@@ -60,6 +59,8 @@
     "@react-native-firebase/app": ">=17.5.0",
     "@react-native-firebase/messaging": ">=17.5.0",
     "@stream-io/react-native-webrtc": ">=104.0.1",
+    "expo": ">=47.0.0",
+    "expo-build-properties": "*",
     "expo-notifications": "*",
     "expo-task-manager": "*",
     "react": ">=17.0.0",
@@ -79,6 +80,12 @@
       "optional": true
     },
     "@react-native-firebase/messaging": {
+      "optional": true
+    },
+    "expo": {
+      "optional": true
+    },
+    "expo-build-properties": {
       "optional": true
     },
     "expo-notifications": {
@@ -115,10 +122,12 @@
     "@tsconfig/node14": "14.1.0",
     "@types/jest": "^29.5.1",
     "@types/lodash.merge": "^4",
-    "@types/react": "18.0.26",
+    "@types/react": "18.2.28",
     "@types/react-native": "^0.70.4",
     "@types/react-native-incall-manager": "^3.2.1",
     "@types/react-test-renderer": "^18",
+    "expo": "49.0.13",
+    "expo-build-properties": "^0.5.1",
     "expo-modules-core": "1.5.11",
     "expo-notifications": "0.20.1",
     "expo-task-manager": "11.3.0",
@@ -132,10 +141,10 @@
     "react-native-svg": "^13.6.0",
     "react-native-voip-push-notification": "3.3.1",
     "react-test-renderer": "^18.2.0",
-    "rimraf": "^5.0.1",
+    "rimraf": "^5.0.5",
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.5"
+    "typescript": "^5.2.2"
   },
   "react-native-builder-bob": {
     "source": "src",
