@@ -1,18 +1,17 @@
 {
   "name": "@stream-io/video-react-sdk",
-  "version": "0.3.40",
+  "version": "0.3.44",
   "packageManager": "yarn@3.2.4",
-  "main": "./dist/index.js",
+  "main": "./dist/index.cjs.js",
+  "module": "./dist/index.es.js",
   "types": "./dist/index.d.ts",
   "license": "See license in LICENSE",
   "scripts": {
-    "postinstall": "yarn copy-version",
     "clean": "rimraf dist",
-    "start": "yarn copy-version && tsc --project tsconfig.json --watch",
-    "build": "yarn copy-version && tsc --project tsconfig.production.json && yarn copy-css",
+    "start": "rollup -c -w",
+    "build": "NODE_ENV=production rollup -c && yarn copy-css",
     "copy-css": "cp -R ../../node_modules/@stream-io/video-styling/dist/* dist/",
-    "start:docs": "npx stream-chat-docusaurus -s",
-    "copy-version": "echo \"export const version = '$npm_package_version';\" > ./version.ts"
+    "start:docs": "npx stream-chat-docusaurus -s"
   },
   "repository": {
     "type": "git",
@@ -32,10 +31,9 @@
     "@floating-ui/react": "^0.22.0",
     "@nivo/core": "^0.80.0",
     "@nivo/line": "^0.80.0",
-    "@stream-io/i18n": "workspace:^",
     "@stream-io/video-client": "workspace:^",
     "@stream-io/video-react-bindings": "workspace:^",
-    "clsx": "^1.2.1",
+    "clsx": "^2.0.0",
     "prop-types": "^15.8.1",
     "rxjs": "~7.8.1"
   },
@@ -44,11 +42,15 @@
     "react-dom": "^18.0.0"
   },
   "devDependencies": {
+    "@rollup/plugin-json": "^6.0.1",
+    "@rollup/plugin-replace": "^5.0.4",
+    "@rollup/plugin-typescript": "^11.1.5",
     "@stream-io/video-styling": "workspace:^",
-    "@types/prop-types": "^15.7.5",
+    "@types/prop-types": "^15.7.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "rimraf": "^5.0.1",
-    "typescript": "^4.9.5"
+    "rimraf": "^5.0.5",
+    "rollup": "^3.29.4",
+    "typescript": "^5.2.2"
   }
 }
