@@ -62,20 +62,20 @@
     "@tsconfig/react-native": "^2.0.2",
     "@types/express": "^4",
     "@types/jest": "^29.2.1",
-    "@types/react": "^18.0.24",
+    "@types/react": "^18.2.28",
     "@types/react-native-dotenv": "^0.2.0",
     "@types/react-native-incall-manager": "^3.2.1",
     "@types/react-native-video": "^5",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.2.1",
-    "eslint": "^8.19.0",
+    "eslint": "^8.51.0",
     "express": "^4.18.2",
     "jest": "^29.2.1",
     "metro-react-native-babel-preset": "0.73.10",
     "patch-package": "^6.4.7",
-    "prettier": "2.8.4",
+    "prettier": "2.8.8",
     "react-test-renderer": "18.2.0",
-    "typescript": "^4.9.5"
+    "typescript": "^5.2.2"
   },
   "jest": {
     "preset": "react-native"
