@@ -41,9 +41,9 @@
     "@rnx-kit/babel-preset-metro-react-native": "1.1.4",
     "@rnx-kit/metro-config": "1.3.9",
     "@rnx-kit/metro-resolver-symlinks": "0.1.32",
-    "@types/react": "~18.2.14",
+    "@types/react": "~18.2.28",
     "@types/react-native-incall-manager": "^4",
-    "typescript": "^5.1.3"
+    "typescript": "^5.2.2"
   },
   "private": true,
   "installConfig": {
