@@ -1,15 +1,17 @@
 import path from 'path';
 import fs from 'fs';
 
-export function getFixture(
-  name: 'AppDelegate.mm' | 'MainApplication.java' | 'AndroidManifest.xml',
-): string {
+type FileName =
+  | 'AppDelegate.mm'
+  | 'MainApplication.java'
+  | 'AndroidManifest.xml'
+  | 'app-build.gradle';
+
+export function getFixture(name: FileName): string {
   const filepath = path.join(__dirname, name);
   return fs.readFileSync(filepath, 'utf8');
 }
 
-export function getFixturePath(
-  name: 'AppDelegate.mm' | 'MainApplication.java' | 'AndroidManifest.xml',
-): string {
+export function getFixturePath(name: FileName): string {
   return path.join(__dirname, name);
 }
